[BH-372] Turned i18n into separate library
Turned i18n module into separate library moving source files, test
files and creating separate CMakeLists.txt, fixed other modules
link targets, fixed paths and sorted some lists
