+++ module-apps/application-onboarding/model/EULARepository.cpp
@@ -1,10 +1,9 @@
 // Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
-#include <gsl/util>
-#include "module-utils/i18n/i18n.hpp"
-
 #include "EULARepository.hpp"
+#include <gsl/util>
+#include <i18n/i18n.hpp>
 #include <fstream>
 #include <sstream>
