+++ module-utils/Utils.hpp
@@ -3,9 +3,7 @@
 
 #pragma once
 #include "Split.hpp"
-#include <i18n/i18n.hpp>
 #include <log.hpp>
-
 #include <magic_enum.hpp>
 
 #include <algorithm> // std::find_if_not
