+++ module-utils/i18n/i18n.cpp
@@ -3,7 +3,6 @@
 
 #include "i18nImpl.hpp"
 #include <log.hpp>
-#include <Utils.hpp>
 #include <cstdio>
 #include <purefs/filesystem_paths.hpp>
